Revert "Exercises 5.5 to 5.10 ready"
This reverts commit 70d0f7494b4ccccef0c2dda2a9b7c734d315cce3.

+++ part4/blog_list/controllers/blog.js
@@ -31,25 +31,25 @@ blogsRouter.post('/', async (req, res) => {
 })
 
 blogsRouter.delete('/:id', async (req, res) => {
-  try {
-    await Blog.findByIdAndRemove(req.params.id)
-    res.status(204).end()
-  } catch {
-    res.status(400).end()
+  const user = await User.findById(req.user)
+  const blog = await Blog.findById(req.params.id)
+  if (blog.user.toString() !== user._id.toString()) {
+    return res.status(401).json({ error: 'token missing or invalid' })
   }
+  await Blog.findByIdAndRemove(req.params.id)
+  res.status(204).end()
   console.log('blog deleted')
 })
 
 blogsRouter.put('/:id', async (req, res) => {
-  console.log('put')
   const body = req.body
-  const blog = {
+  const blog = new Blog({
     title: body.title,
     author: body.author,
     url: body.url,
     likes: body.likes
-  }
-  const postBlog = await Blog.findByIdAndUpdate(req.params.id, blog, { new: true })
+  })
+  const postBlog = await blog.save()
   res.status(201)
   res.json(postBlog)
 })
